+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
  ## [Unreleased]
 
+## [2.4.0] - 2023-
+### Added
+- Add Webhook `ValidateRequest`
+
 ## [2.3.4] - 2020-09-09
 ### Fixed
 - Correctly ignore non-call events and clear handlers on call end